Skip cors protection for rails 4.1
In rails 4.1 cors protection is more strict. But since
with netzke we need to generate the js file on the
fly we need to skip the verification altogether.

A better solution would be to somehow integrate the 
dynamic ext.css and ext.js-file into the asset
pipeline, and let it build by that. Normally it
does not change often, except during development.